requirements.txt
# Flask核心框架
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-Caching==2.1.0
Flask-Limiter==3.5.0

# 数据库
SQLAlchemy==2.0.21
PyMySQL==1.1.0
cryptography==41.0.4

# 表单处理
WTForms==3.0.1
WTForms-JSON==0.3.5

# 密码加密
bcrypt==4.0.1
passlib==1.7.4

# JWT认证
PyJWT==2.8.0
cryptography==41.0.4

# 图像处理和人脸识别
Pillow==10.0.1
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3

# 数据处理和可视化
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# 日期时间处理
python-dateutil==2.8.2
pytz==2023.3

# 文件处理
openpyxl==3.1.2
xlrd==2.0.1
python-magic==0.4.27

# HTTP请求
requests==2.31.0
urllib3==2.0.4

# 配置和环境变量管理
python-dotenv==1.0.0
marshmallow==3.20.1

# 任务队列
celery==5.3.2
redis==5.0.0

# 邮件服务
email-validator==2.0.0

# 日志处理
loguru==0.7.2

# 开发工具
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0
isort==5.12.0

# 生产环境
gunicorn==21.2.0
eventlet==0.33.3

# 系统监控
psutil==5.9.5

# 其他工具
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.7